+++ lib/SimpleDataTable.jsx
@@ -1,8 +1,8 @@
 // Import modules
 import React from 'react'
 import PropTypes from 'prop-types'
-import Table from './Table.jsx'
-import Link from './Link.jsx'
+import DefaultTable from './DefaultTable'
+import Link from './Link'
 import { getNestedValue, getLastValue } from './helper-functions'
 
 /*
@@ -29,7 +29,7 @@ import { getNestedValue, getLastValue } from './helper-functions'
 
 const renderCellFromHeader = ({
   path, link, transform, appendValue,
-}, row) => {
+}, row, Table) => {
   const key = getLastValue(path)
   const isLink = typeof link === 'string'
   const value = getNestedValue(row, path)
@@ -51,8 +51,19 @@ const renderCellFromHeader = ({
 
 const SimpleDataTable = (props) => {
   const {
-    keyId, headers, data, vertical, rightAlignedNames, ...otherProps
+    keyId, headers, data, vertical, rightAlignedNames, customElements,
+    ...otherProps
   } = props
+  const {
+    table, tbody, td, tfoot, th, thead, tr,
+  } = customElements
+  const Table = table || DefaultTable
+  Table.Body = tbody || DefaultTable.Body
+  Table.Cell = td || DefaultTable.Cell
+  Table.Footer = tfoot || DefaultTable.Footer
+  Table.HeaderCell = th || DefaultTable.HeaderCell
+  Table.Header = thead || DefaultTable.Header
+  Table.Row = tr || DefaultTable.Row
   return vertical ? (
     <Table {...otherProps}>
       <Table.Body>
@@ -63,7 +74,7 @@ const SimpleDataTable = (props) => {
                 <Table.Cell header rightAlignedNames={rightAlignedNames}>
                   {header.name}
                 </Table.Cell>
-                {renderCellFromHeader(header, row)}
+                {renderCellFromHeader(header, row, Table)}
               </Table.Row>
             ))}
           </>
@@ -84,7 +95,7 @@ const SimpleDataTable = (props) => {
       <Table.Body>
         {data.map((row, idx) => (
           <Table.Row key={keyId ? row[keyId] : idx}>
-            {headers.map(header => renderCellFromHeader(header, row))}
+            {headers.map(header => renderCellFromHeader(header, row, Table))}
           </Table.Row>
         ))}
       </Table.Body>
@@ -99,6 +110,15 @@ SimpleDataTable.propTypes = {
   headers: PropTypes.arrayOf(PropTypes.object).isRequired,
   vertical: PropTypes.bool,
   rightAlignedNames: PropTypes.bool,
+  customElements: PropTypes.exact({
+    table: PropTypes.element,
+    tbody: PropTypes.element,
+    td: PropTypes.element,
+    tfoot: PropTypes.element,
+    th: PropTypes.element,
+    thead: PropTypes.element,
+    tr: PropTypes.element,
+  }),
 }
 
 // Defines the default values for not passing a certain prop
@@ -106,6 +126,7 @@ SimpleDataTable.defaultProps = {
   keyId: '',
   vertical: false,
   rightAlignedNames: false,
+  customElements: {},
 }
 
 export default SimpleDataTable